feat: enhance BaseController with configurable options for endpoint support and improved error handling

+++ src/common/base.controller.ts
@@ -1,4 +1,4 @@
-import { Delete, Get, HttpCode, HttpStatus, Inject, Param, ParseUUIDPipe, Query, Type, Version } from '@nestjs/common';
+import { Delete, Get, HttpCode, HttpStatus, Inject, NotFoundException, Param, ParseUUIDPipe, Query, Type, Version } from '@nestjs/common';
 import { REQUEST } from '@nestjs/core';
 import { ApiOperation } from '@nestjs/swagger';
 import { Request as ExpressRequest } from 'express';
@@ -6,7 +6,7 @@ import { Request as ExpressRequest } from 'express';
 import { PaginateConfig, PaginateQuery, Paginated } from 'nestjs-paginate';
 import { BaseApiOkResponse } from '../decorators/base-response.decorator';
 import { PayloadDto } from './dto/payload.dto';
-import { IBaseController } from './interfaces/base-controller.interface';
+import { BaseControllerOptions, IBaseController } from './interfaces/base-controller.interface';
 import { IBaseService } from './interfaces/base-service.interface';
 
 export class BaseController<T, D> implements IBaseController<T, D> {
@@ -15,9 +15,18 @@ export class BaseController<T, D> implements IBaseController<T, D> {
     protected service: IBaseService<T, D>;
     protected configPagination: PaginateConfig<T>;
 
-    constructor(service: IBaseService<T, D>, configPagination?: PaginateConfig<T>) {
+    private readonly options: Required<BaseControllerOptions>;
+
+    constructor(service: IBaseService<T, D>, configPagination?: PaginateConfig<T>, options?: BaseControllerOptions) {
         this.service = service;
         this.configPagination = configPagination;
+
+        this.options = {
+            enableDelete: options?.enableDelete ?? true,
+            enableGetAll: options?.enableGetAll ?? true,
+            enableGetById: options?.enableGetById ?? true,
+            enablePagination: options?.enablePagination ?? true,
+        };
     }
 
     protected getRequest(): PayloadDto | null {
@@ -33,6 +42,10 @@ export class BaseController<T, D> implements IBaseController<T, D> {
     @Get('all')
     @BaseApiOkResponse(Object as unknown as Type<D>, { isArray: true })
     async getAll(): Promise<D[]> {
+        if (!this.options.enableGetAll) {
+            throw new NotFoundException('Endpoint not supported');
+        }
+
         const result = await this.service.findAll();
         return result;
     }
@@ -43,6 +56,10 @@ export class BaseController<T, D> implements IBaseController<T, D> {
     @Get(':id')
     @BaseApiOkResponse(Object as unknown as Type<D>)
     async getById(@Param('id', new ParseUUIDPipe()) id: string): Promise<D> {
+        if (!this.options.enableGetById) {
+            throw new NotFoundException('Endpoint not supported');
+        }
+
         const result = await this.service.findById(id);
         return result;
     }
@@ -53,6 +70,9 @@ export class BaseController<T, D> implements IBaseController<T, D> {
     @Get()
     @BaseApiOkResponse(Object as unknown as Type<Paginated<D>>)
     async getPagination(@Query() query: PaginateQuery): Promise<Paginated<D>> {
+        if (!this.options.enablePagination) {
+            throw new NotFoundException('Endpoint not supported');
+        }
         const result = await this.service.getPaginationWithCustomQuery(query, this.configPagination);
         return result;
     }
@@ -63,6 +83,10 @@ export class BaseController<T, D> implements IBaseController<T, D> {
     @Delete(':id')
     @BaseApiOkResponse(Boolean)
     async delete(@Param('id', new ParseUUIDPipe()) id: string): Promise<boolean> {
+        if (!this.options.enableDelete) {
+            throw new NotFoundException('Endpoint not supported');
+        }
+
         const result = await this.service.delete(id);
         return result;
     }
